Updated Notifications Framework

+++ frontend/components/InventoryItem.js
@@ -1,50 +1,59 @@
 import React, { useState } from "react";
-import { View, StyleSheet, TouchableOpacity} from 'react-native'
-import { Text, Image} from "react-native-elements";
+import { View, StyleSheet, TouchableOpacity } from "react-native";
+import { Text, Image } from "react-native-elements";
 import { useItems } from "../providers/ItemsProvider";
 import { ActionSheet } from "./ActionSheet";
 import { Item } from "../schemas";
-
 import styles from "../stylesheet";
 
-export function InventoryItem({ item }) {
+export function InventoryItem({ item, name }) {
   const [actionSheetVisible, setActionSheetVisible] = useState(false);
   const { deleteItem, setItemStatus } = useItems();
-  const actions = [
-    {
-      title: "Delete",
-      action: () => {
-        deleteItem(item);
-      },
-    },
-  ];
+  const actions = [];
 
   const itemStatus = (item) => {
-
-    const status = item.status
-    if(status === Item.STATUS_COMPLETE){
-      return <Text style={myStyles.subText}>Status: Out of Stock</Text>
-    } else if (status === Item.STATUS_IN_PROGRESS){
-      return <Text style={myStyles.subText}>Status: Running Low</Text>
-    } else if (status === Item.STATUS_OPEN){
-      return <Text style={myStyles.subText}>Status: Stocked</Text>
+    const status = item.status;
+    if (status === Item.STATUS_COMPLETE) {
+      return <Text style={myStyles.subText}>Status: Out of Stock</Text>;
+    } else if (status === Item.STATUS_IN_PROGRESS) {
+      return <Text style={myStyles.subText}>Status: Running Low</Text>;
+    } else if (status === Item.STATUS_OPEN) {
+      return <Text style={myStyles.subText}>Status: Stocked</Text>;
     } else {
-      return null
+      return null;
     }
-  }
+  };
 
   const itemImage = (item) => {
-    if(item.image !== ""){
-      return <Image source={{uri:`${item.image}`}} style={{width:100, height:100}}/>
+    if (item.image !== "") {
+      return (
+        <Image
+          source={{ uri: `${item.image}` }}
+          style={{ width: 100, height: 100 }}
+        />
+      );
     } else {
-      return <Image source={require("../assets/img/cart.jpg")} style={{width:100,height:100}}/>
+      return (
+        <Image
+          source={require("../assets/img/cart.jpg")}
+          style={{ width: 100, height: 100 }}
+        />
+      );
     }
-    
-  }
+  };
   // For each possible status other than the current status, make an action to
   // move the item into that status. Rather than creating a generic method to
   // avoid repetition, we split each status to separate each case in the code
   // below for demonstration purposes.
+  if (name === "My Inventory") {
+    actions.push({
+      title: "Delete",
+      action: () => {
+        deleteItem(item);
+      },
+    });
+  }
+
   if (item.status !== "" && item.status !== Item.STATUS_OPEN) {
     actions.push({
       title: "Fully Stocked",
@@ -72,7 +81,7 @@ export function InventoryItem({ item }) {
 
   return (
     <>
-    {/* Do not mess with ActionSheet */}
+      {/* Do not mess with ActionSheet */}
       <ActionSheet
         visible={actionSheetVisible}
         closeOverlay={() => {
@@ -83,62 +92,48 @@ export function InventoryItem({ item }) {
         actions={actions}
       />
 
-      
-      
-        {/* <Image source={require("../assets/img/cart.jpg")} style={{width:150,height:150}}/> */}
-          <TouchableOpacity style={[myStyles.item, styles.navBarShadow]}
-            onPress={() => {setActionSheetVisible(true);}}
-          
-          >
-            <View >
-              {/* <Image source={{uri:`${item.image}`}} style={{width:100,height:100}}/> */}
-              {itemImage(item)}
-            </View>
+      {/* <Image source={require("../assets/img/cart.jpg")} style={{width:150,height:150}}/> */}
+      <TouchableOpacity
+        style={[myStyles.item, styles.navBarShadow]}
+        onPress={() => {
+          setActionSheetVisible(true);
+        }}
+      >
+        <View>
+          {/* <Image source={{uri:`${item.image}`}} style={{width:100,height:100}}/> */}
+          {itemImage(item)}
+        </View>
 
-            <View style={{flex: 2}}>
-              <View style={{marginLeft:10}}>
-                <Text style={myStyles.itemName}>{item.name}</Text>
-                
-                <Text style = {myStyles.subText}> 
-                  Quantity: {item.quantity}
-                </Text>
-                <Text style = {myStyles.subText}> 
-                  Brand: {item.brand}
-                </Text>
-                <Text style = {myStyles.subText}> 
-                  Purchase Date: {item.date}
-                </Text>
+        <View style={{ flex: 2 }}>
+          <View style={{ marginLeft: 10 }}>
+            <Text style={myStyles.itemName}>{item.name}</Text>
 
-                {itemStatus(item)}
+            <Text style={myStyles.subText}>Quantity: {item.quantity}</Text>
+            <Text style={myStyles.subText}>Brand: {item.brand}</Text>
+            <Text style={myStyles.subText}>Purchase Date: {item.date}</Text>
 
-              </View>
-            </View>
-
-            {/* <View>
-              {itemStatus(item)}
-            </View> */}
-          </TouchableOpacity>
-        
-        
-      
+            {itemStatus(item)}
+          </View>
+        </View>
+      </TouchableOpacity>
     </>
   );
 }
 
 const myStyles = StyleSheet.create({
-  item : {
-    flexDirection: 'row',
-    alignItems: 'center',
-    backgroundColor: 'white',
+  item: {
+    flexDirection: "row",
+    alignItems: "center",
+    backgroundColor: "white",
     marginVertical: 3,
     padding: 10,
-    alignContent:'space-around'
+    alignContent: "space-around",
   },
   itemName: {
-    fontWeight: 'bold',
-    fontSize: 15
+    fontWeight: "bold",
+    fontSize: 15,
   },
   subText: {
-    fontSize: 12
-  }
+    fontSize: 12,
+  },
 });
